requirements.txt
PyPDF2==3.0.1
Flask==3.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
google-cloud-storage>=2.0.0
Werkzeug>=3.0.0
